#g-f-a-i-p Modified the binarisation parameters

+++ src/gui/image_process.h
@@ -10,20 +10,30 @@ typedef enum { MEAN,
 			STD_DEVIATION } II_TYPE;
 
 void init_sdl();
+
 SDL_Surface* load_image(char *path);
+
 void color_surface(SDL_Surface *surface, Uint32 pixel);
-unsigned long int **create_integral_image(SDL_Surface *surface, 
+
+unsigned long long int **create_integral_image(SDL_Surface *surface, 
 					unsigned int margin, II_TYPE ii_type);
-void free_integral_image(unsigned long int **integral_image, size_t width);
-unsigned long int integral_image_value(II_TYPE ii_type, Uint8 composant);
+
+void free_integral_image(unsigned long long int **integral_image, size_t width);
+
+unsigned long long int integral_image_value(II_TYPE ii_type, Uint8 composant);
+
 void grayscale(Uint8 *r, Uint8 *g, Uint8 *b);
+
 void gaussian_blur(SDL_Surface *src_surface, int i, int j, 
 	Uint8 *r_dest, Uint8 *g_dest, Uint8 *b_dest);
-float sauvola_binarisation(SDL_Surface *src_surface, 
-						unsigned long int **mean_ii, 
-						unsigned long int **std_deviation_ii, 
+
+double sauvola_binarisation(SDL_Surface *src_surface, 
+						unsigned long long int **mean_ii, 
+						unsigned long long int **std_deviation_ii, 
 						int i, int j);
+
 void image_process(char *path);
+
 void rotate_image(char *path, double angle);
 
 #endif
